refactor(naming): rename util.ts to utils.ts
